Create PHBClasses_Orig_LW
This is the original coding for the classes as provided in the new help file that Lone Wolf provided. 

Some lines of this script would not compile. They have been commented out. 

I had already updated all spell references that were not included in the SRD to point to our new spell names.