Rename PHBSpells-Public.user to backup/PHBSpells-Public.user
Moving this to backup. Based on the feedback from LW we should be ok to use spell names. All descriptive text has been removed or paraphrased from the original file. A backup has also been made of the real original. Keeping this in backup in case we need to revert to new spell names again in the future.