Update COM_5ePack_PHB - Feats.user
Thought this might be linking to the Energy Type options in COM_5ePack - Tags.1st. 
Does not appear to be the case. This leads me to believe this is not an issue with the listing in the Community Pack?
Energy Type (Base 4) does not list fire. 
I can find no mechanics as a result of this choice so have changed to an array-based menu with all relevant options.  
Tested OK.