Update COM_5ePack_UA - Barbarian.user
I completely altered the implementation of the Path of the Storm Herald.  I made it better reflect the form and function of Path of the Totem Warrior by implementing a configuration where you choose the environment.  all abilities granted by the path of the storm herald now query which environment was chosen and automatically update their livename and CustDesc.