#542 - Fixed Ranger (Great Favored Enemy and Favored Enemy)
Simplified the Favored Enemy and split it into 2x abilities instead of 1. When the UA changed ranger is removed the requirement to have a complex coded solution. 

ISSUE: If you pick Humanoid as a language choice you get 2x languages instead of 1! It's the same code so I'm sure this is coming from somewhere else!