Artificer
Implements the new Artificer class described in
http://dnd.wizards.com/articles/unearthed-arcana/artificer

Notes:
The Mechanical Servant is not turned into a Construct due to a
validation error, since the Racial tag expression demands the eligible
creatures to be Beasts.
The Mechanical Servant does not gain it's master languages, since I
could not check which ones he gained after the level 6, as said by the
article.
The Artificer Spell List was implemented using an ugly rack on the tag
expressions govering the available spells. This was done because I did
not want to duplicate every single spell in the list just to add the
sClass.cHelpArt tag to them. This might bite me in the future...