Update COM_5ePack_MTOF - Monsters.user
This is the N monsters. Before I dive into the next batch, I'd love it if someone can take a look. There's one ability that's listed as incomplete, as I couldn't figure out how to bolt on the two damage types. But I wanted to get the rest up for review.