UA Gothic - Replaced "Tools" skill
This was programmed with a new "Tools" skill for the sake of helping program the Monster Hunter Fighter's Bonus Proficiencies feature. But if the UA is selected as a source, that skill shows up for everyone. I replaced the code with the Simple Thing called Tools. The functionality is the same, but without the unnecessary and confusing "Tools" skill ever showing up; it's deleted.